requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
pytest
httpx
psycopg2-binary
